helloYouProj/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
daphne==3.0.2
Django==4.1.6
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
drf-yasg==1.21.5
GDAL==3.6.2
hyperlink==21.0.0
idna==3.4
importlib-resources==5.12.0
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
msgpack==1.0.4
openapi-codec==1.3.2
packaging==23.0
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
simplejson==3.18.3
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.14.0
zope.interface==5.5.2
fontawesomefree==6.4.0